[Feature] docs, service interface 추가
