+++ src/main/java/com/greeni/api/activities/docs/ActivityControllerDocs.java
@@ -1,0 +1,7 @@
+package com.greeni.api.activities.docs;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+@Tag(name = "Activity", description = "활동내역 CR API")
+public interface ActivityControllerDocs {
+}
