+++ src/main/java/com/greeni/api/activities/service/ActivityService.java
@@ -1,7 +1,4 @@
 package com.greeni.api.activities.service;
 
-import org.springframework.stereotype.Service;
-
-@Service
-public class ActivityService {
+public interface ActivityService {
 }
