+++ src/main/java/com/greeni/api/badges/controller/BadgeController.java
@@ -3,7 +3,14 @@ package com.greeni.api.badges.controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.greeni.api.badges.docs.BadgeControllerDocs;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
 @RestController
 @RequestMapping("/api/badges")
-public class BadgeController {
+@RequiredArgsConstructor
+public class BadgeController implements BadgeControllerDocs {
 }
