+++ src/main/java/com/greeni/api/badges/docs/BadgeControllerDocs.java
@@ -1,0 +1,7 @@
+package com.greeni.api.badges.docs;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+@Tag(name = "Badge", description = "배지 CRUD API")
+public interface BadgeControllerDocs {
+}
