+++ src/main/java/com/greeni/api/badges/service/BadgeService.java
@@ -1,7 +1,4 @@
 package com.greeni.api.badges.service;
 
-import org.springframework.stereotype.Service;
-
-@Service
-public class BadgeService {
+public interface BadgeService {
 }
