+++ src/main/java/com/greeni/api/badges/service/BadgeServiceImpl.java
@@ -1,0 +1,14 @@
+package com.greeni.api.badges.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+@Transactional
+public class BadgeServiceImpl implements BadgeService {
+}
