+++ src/main/java/com/greeni/api/diaries/docs/DiaryControllerDocs.java
@@ -1,0 +1,7 @@
+package com.greeni.api.diaries.docs;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+@Tag(name = "Diary", description = "일기 CRU API")
+public interface DiaryControllerDocs {
+}
