+++ src/main/java/com/greeni/api/diaries/service/DiaryService.java
@@ -1,9 +1,4 @@
 package com.greeni.api.diaries.service;
 
-import jakarta.transaction.Transactional;
-import org.springframework.stereotype.Service;
-
-@Service
-@Transactional
-public class DiaryService {
+public interface DiaryService {
 }
