+++ src/main/java/com/greeni/api/diaries/service/DiaryServiceImpl.java
@@ -1,0 +1,14 @@
+package com.greeni.api.diaries.service;
+
+import org.springframework.stereotype.Service;
+
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class DiaryServiceImpl implements DiaryService {
+}
