+++ src/main/java/com/greeni/api/members/controller/MemberController.java
@@ -1,47 +1,53 @@
 package com.greeni.api.members.controller;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import com.greeni.api.apiPayload.CommonResponse;
+import com.greeni.api.members.docs.MemberControllerDocs;
 import com.greeni.api.members.dto.MemberRequestDTO;
 import com.greeni.api.members.dto.MemberResponseDTO;
 import com.greeni.api.members.service.MemberService;
-import io.swagger.v3.oas.annotations.Operation;
+
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @RestController
 @RequestMapping("/api/members")
 @RequiredArgsConstructor
-public class MemberController {
+public class MemberController implements MemberControllerDocs {
 
-    private final MemberService memberService;
+	private final MemberService memberService;
 
-    @PostMapping("/signup")
-    @Operation(summary = "회원가입 API")
-    public ResponseEntity<CommonResponse<MemberResponseDTO.toMemberResultDTO>> signUp(@RequestBody @Valid MemberRequestDTO.SignUpDTO request){
-        MemberResponseDTO.toMemberResultDTO result = memberService.signUpMember(request);
-        return new ResponseEntity<>(CommonResponse.created(result),HttpStatus.CREATED);
-    }
+	@PostMapping("/signup")
+	public ResponseEntity<CommonResponse<MemberResponseDTO.toMemberResultDTO>> signUp(
+		@RequestBody @Valid MemberRequestDTO.SignUpDTO request) {
+		MemberResponseDTO.toMemberResultDTO result = memberService.signUpMember(request);
+		return new ResponseEntity<>(CommonResponse.created(result), HttpStatus.CREATED);
+	}
 
-    @PostMapping("/email")
-    @Operation(summary = "이메일 인증번호 전송 API")
-    public ResponseEntity<CommonResponse<?>> emailRequest(@RequestBody @Valid MemberRequestDTO.EmailDTO request){
-        memberService.sendEmail(request);
-        return ResponseEntity.ok().body(CommonResponse.onSuccess(null));
-    }
+	@PostMapping("/email")
+	public ResponseEntity<CommonResponse<?>> emailRequest(@RequestBody @Valid MemberRequestDTO.EmailDTO request) {
+		memberService.sendEmail(request);
+		return ResponseEntity.ok().body(CommonResponse.onSuccess(null));
+	}
 
-    @PostMapping("/password")
-    @Operation(summary = "비밀번호 찾기 API")
-    public ResponseEntity<CommonResponse<MemberResponseDTO.toMemberResultDTO>> findPassword(@RequestBody @Valid MemberRequestDTO.PasswdDTO request){
-        return ResponseEntity.ok().body(CommonResponse.onSuccess(memberService.findPw(request)));
-    }
+	@PostMapping("/password")
+	public ResponseEntity<CommonResponse<MemberResponseDTO.toMemberResultDTO>> findPassword(
+		@RequestBody @Valid MemberRequestDTO.PasswdDTO request) {
+		return ResponseEntity.ok().body(CommonResponse.onSuccess(memberService.findPw(request)));
+	}
 
-    @PostMapping("/password/reset")
-    @Operation(summary = "비밀번호 재설정 API")
-    public ResponseEntity<CommonResponse<MemberResponseDTO.toMemberResultDTO>> setPassword(@RequestBody @Valid MemberRequestDTO.ResetPwDTO request){
-        return ResponseEntity.ok().body(CommonResponse.onSuccess(memberService.resetPw(request)));
-    }
+	@PostMapping("/password/reset")
+	public ResponseEntity<CommonResponse<MemberResponseDTO.toMemberResultDTO>> setPassword(
+		@RequestBody @Valid MemberRequestDTO.ResetPwDTO request) {
+		return ResponseEntity.ok().body(CommonResponse.onSuccess(memberService.resetPw(request)));
+	}
 
 }
