+++ src/main/java/com/greeni/api/members/service/MemberService.java
@@ -1,151 +1,20 @@
 package com.greeni.api.members.service;
 
-import java.io.UnsupportedEncodingException;
-import java.util.concurrent.TimeUnit;
-
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.ValueOperations;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import com.greeni.api.apiPayload.exception.GeneralException;
-import com.greeni.api.apiPayload.status.MemberErrorStatus;
-import com.greeni.api.members.converter.MemberConverter;
 import com.greeni.api.members.domain.Member;
 import com.greeni.api.members.dto.MemberRequestDTO;
 import com.greeni.api.members.dto.MemberResponseDTO;
-import com.greeni.api.members.repository.MemberRepository;
 
-import jakarta.mail.MessagingException;
-import jakarta.mail.internet.InternetAddress;
-import jakarta.mail.internet.MimeMessage;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+public interface MemberService {
 
-@Service
-@Transactional
-@Slf4j
-@RequiredArgsConstructor
-public class MemberService {
+	MemberResponseDTO.toMemberResultDTO signUpMember(MemberRequestDTO.SignUpDTO request);
 
-	private final MemberRepository memberRepository;
-	private final JavaMailSender javaMailSender;
-	private final RedisTemplate<String, Object> redistemplate;
-	private final PasswordEncoder passwordEncoder;
+	void checkEmailCode(String email, String code);
 
-	public MemberResponseDTO.toMemberResultDTO signUpMember(MemberRequestDTO.SignUpDTO request) {
-		// 해당 메일이 이미 존재하는 메일인지 확인
-		if (memberRepository.existsByEmail(request.getEmail())) {
-			throw new GeneralException(MemberErrorStatus.EXIST_EMAIL);
-		}
+	void sendEmail(MemberRequestDTO.EmailDTO request);
 
-		// 이메일 인증 완료 여부 확인
-		checkEmailCode(request.getEmail(), request.getCode());
-		//        ValueOperations<String, Object> ops = redistemplate.opsForValue();
-		//        String codeNum = (String) ops.get("EmailCode"+request.getEmail());
-		//        if(codeNum == null){
-		//            throw new GeneralException(ErrorStatus.EXPIRED_CODE);
-		//        }
-		//        if(!codeNum.equals(request.getCode())){
-		//            throw new GeneralException(ErrorStatus.WRONG_CODE);
-		//        }
+	MemberResponseDTO.toMemberResultDTO findPw(MemberRequestDTO.PasswdDTO request);
 
-		Member member = MemberConverter.toMember(request);
-		member.encodePassword(passwordEncoder.encode(member.getPassword()));
-		memberRepository.save(member);
+	MemberResponseDTO.toMemberResultDTO resetPw(MemberRequestDTO.ResetPwDTO request);
 
-		return MemberConverter.toMemberResultDTO(member);
-	}
-
-	public void checkEmailCode(String email, String code) {
-		ValueOperations<String, Object> ops = redistemplate.opsForValue();
-		String codeNum = (String)ops.get("EmailCode" + email);
-		if (codeNum == null) {
-			throw new GeneralException(MemberErrorStatus.EXPIRED_CODE);
-		}
-		if (!codeNum.equals(code)) {
-			throw new GeneralException(MemberErrorStatus.WRONG_CODE);
-		}
-	}
-
-	// 이메일 인증 번호 전송
-	public void sendEmail(MemberRequestDTO.EmailDTO request) {
-		int number = (int)(Math.random() * 90000) + 100000;
-
-		try {
-			MimeMessage message = javaMailSender.createMimeMessage();
-
-			MimeMessageHelper helper = new MimeMessageHelper(message, true, "UTF-8");
-			helper.setSubject("[Greeni] 이메일 인증 번호");
-			helper.setTo(request.getEmail());
-			helper.setFrom(new InternetAddress("greeni@greeni.com", "greeni", "UTF-8"));
-
-			String body = "";
-			body += "<!DOCTYPE html>";
-			body += "<html lang='ko'>";
-			body += "<head><meta charset='UTF-8'></head>";
-			body += "<body style='margin:0;padding:0;background-color:#f6f7f8;font-family:sans-serif;'>";
-
-			body += "<div style='width:100%;padding:30px 0;'>";
-			body += "  <div style='max-width:600px;margin:0 auto;background:#ffffff;"
-				+ "border-radius:8px;padding:30px;box-shadow:0 4px 12px rgba(0,0,0,0.08);'>";
-
-			body += "    <h3 style='margin-top:0;color:#333;'>Greeni의 인증번호입니다.</h3>";
-			body += "    <h1 style='font-size:32px;color:#2d3748;text-align:center;margin:20px 0;'>"
-				+ number + "</h1>";
-			body += "    <p style='font-size:15px;color:#555;line-height:1.6;'>"
-				+ "화면으로 돌아가 인증번호를 입력해주세요<br/>"
-				+ "인증번호는 발송된 시점부터 <b>3분</b> 동안 유효합니다.<br/>"
-				+ "유효시간 내 인증을 완료하지 않을 경우 재요청이 필요합니다."
-				+ "</p>";
-
-			body += "  </div>";
-			body += "</div>";
-
-			body += "</body></html>";
-
-			helper.setText(body, true);
-
-			javaMailSender.send(message);
-		} catch (MessagingException | UnsupportedEncodingException e) {
-			log.error("Error Sending email", e);
-			throw new GeneralException(MemberErrorStatus.NOT_SEND_EMAIL_CODE);
-		}
-
-		// redis에 인증번호 3분간 저장
-		ValueOperations<String, Object> ops = redistemplate.opsForValue();
-		ops.set("EmailCode" + request.getEmail(), number + "", 200, TimeUnit.SECONDS);
-	}
-
-	public MemberResponseDTO.toMemberResultDTO findPw(MemberRequestDTO.PasswdDTO request) {
-		Member member = findMemberByEmail(request.getEmail());
-		// 이메일 인증번호 맞는지 확인
-		checkEmailCode(request.getEmail(), request.getCode());
-		// redis에 비밀번호 재설정함을 저장하기
-		ValueOperations<String, Object> ops = redistemplate.opsForValue();
-		ops.set("chPw" + request.getEmail(), "chPw", 300, TimeUnit.SECONDS);
-		return MemberConverter.toMemberResultDTO(member);
-	}
-
-	public MemberResponseDTO.toMemberResultDTO resetPw(MemberRequestDTO.ResetPwDTO request) {
-		Member member = findMemberByEmail(request.getEmail());
-		// 비밀번호 찾기를 했는지 확인
-		ValueOperations<String, Object> ops = redistemplate.opsForValue();
-		String state = (String)ops.get("chPw" + request.getEmail());
-		if (state == null) {
-			throw new GeneralException(MemberErrorStatus.GO_TO_FINDPW);
-		}
-		// 새 비밀번호 암호화해서 저장하기
-		member.encodePassword(passwordEncoder.encode(request.getPassword()));
-		memberRepository.save(member);
-		return MemberConverter.toMemberResultDTO(member);
-	}
-
-	public Member findMemberByEmail(String email) {
-		return memberRepository.findByEmail(email)
-			.orElseThrow(() -> new GeneralException(MemberErrorStatus.NOT_EXIST_EMAIL));
-	}
+	Member findMemberByEmail(String email);
 }
