+++ src/main/java/com/greeni/api/profiles/controller/ProfileController.java
@@ -3,7 +3,14 @@ package com.greeni.api.profiles.controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.greeni.api.profiles.docs.ProfileControllerDocs;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
 @RestController
 @RequestMapping("/api/profiles")
-public class ProfileController {
+@RequiredArgsConstructor
+public class ProfileController implements ProfileControllerDocs {
 }
