+++ src/main/java/com/greeni/api/profiles/docs/ProfileControllerDocs.java
@@ -1,0 +1,7 @@
+package com.greeni.api.profiles.docs;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+@Tag(name = "Profile", description = "프로필 CRUD API")
+public interface ProfileControllerDocs {
+}
