+++ src/main/java/com/greeni/api/profiles/service/ProfileService.java
@@ -1,4 +1,4 @@
 package com.greeni.api.profiles.service;
 
-public class ProfileService {
+public interface ProfileService {
 }
