+++ src/main/java/com/greeni/api/profiles/service/ProfileServiceImpl.java
@@ -1,0 +1,14 @@
+package com.greeni.api.profiles.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+@Transactional
+public class ProfileServiceImpl implements ProfileService {
+}
